+++ Sample16Dlg.h
@@ -8,7 +8,7 @@
 typedef struct ADSDYNSYM_SUBINFO
 {
     CAdsSymbolInfo infoParent;
-    long nSub;
+    uint32_t nSub;
     ADSDYNSYM_SUBINFO* pParent;
 } ADSDYNSYM_SUBINFO, *PADSDYNSYM_SUBINFO;
 
@@ -21,7 +21,7 @@ class CSample16Dlg : public QDialog
     Q_OBJECT
 
 public:
-    explicit CSample16Dlg(const char * targetIp, const AmsNetId & targetNetId, quint16 targetPort, QWidget * parent = nullptr);
+    explicit CSample16Dlg(const char * targetIp, const AmsNetId & targetNetId, uint16_t targetPort, QWidget * parent = nullptr);
     ~CSample16Dlg();
 
 private slots:
@@ -33,28 +33,28 @@ private slots:
 
 private:
     bool initDialog();
-    void handleNavigation(long navType, const QString& errorMessage);
+    void handleNavigation(uint32_t navType, const QString & errorMessage);
 
     AdsDevice m_adsDevice;
 
     QString m_strName;
     QString m_strFullName;
     QString m_strComment;
-    long m_nIndexGroup;
-    long m_nIndexOffset;
-    long m_nSize;
+    uint32_t m_nIndexGroup;
+    uint32_t m_nIndexOffset;
+    uint32_t m_nSize;
     QString m_strType;
-    long m_nAdsType;
+    uint32_t m_nAdsType;
 
     CAdsParseSymbols* m_pDynSymbols;
-    long m_nCurDynSymbol;
+    uint32_t m_nCurDynSymbol;
     ADSDYNSYM_SUBINFO* m_pCurSubSymbol;
-    long m_nNextNavType;
+    uint32_t m_nNextNavType;
 
     Ui::Sample16Dialog * m_ui;
 
-    long adsSetFirstDynSymbol(bool bForceReload);
-    long adsGetNextDynSymbol(long navType, QString& strName, QString& strFullName,
-                             QString& strType, QString& strComment, long& adsType,
-                             long& cbSymbolSize, long& nIndexGroup, long& IndexOffset);
+    uint32_t adsSetFirstDynSymbol(bool bForceReload);
+    uint32_t adsGetNextDynSymbol(uint32_t navType, QString & strName, QString & strFullName,
+                             QString & strType, QString & strComment, uint32_t & adsType,
+                             uint32_t & cbSymbolSize, uint32_t & nIndexGroup, uint32_t & IndexOffset);
 };
